pyproject.toml
[tool.poetry]
name = "flo-gateway"
version = "0.0.1"
description = "A easy way to create structured AI agents"
authors = ["vizsatiz <vishnu@rootfor.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.1.22"
langchain-openai = "^0.1.21"
langchain-community = "^0.2.11"
langchain = "^0.2.12"
openai = "^1.40.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
